# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Latihan Repo Pattern API",
        "description": "Dokumentasi API untuk Latihan Repository Pattern",
        "contact": {
            "email": "aldyimam03@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/student": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Ambil semua data student",
                "operationId": "6ab0dd8497034822dd4c374d026e7693",
                "responses": {
                    "200": {
                        "description": "List of students"
                    },
                    "500": {
                        "description": "Failed to retrieve students"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Membuat data student baru",
                "operationId": "49c17c59519e96b76193622e4f16edea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age",
                                    "room_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Aldy"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 22
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created successfully"
                    },
                    "500": {
                        "description": "Failed to create student"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/student/{id}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Ambil data student berdasarkan ID",
                "operationId": "8d82708d77327b3b783894a9f208f8e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student data"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Student"
                ],
                "summary": "Update data student",
                "operationId": "6d26c078910df68fb02556bad824d34f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Aldy Updated"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 23
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student updated successfully"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Student"
                ],
                "summary": "Delete a student",
                "description": "Delete a student by ID. Only accessible by admin.",
                "operationId": "50eb92799be93d54433e29162b9dc146",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the student",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student deleted successfully"
                    },
                    "404": {
                        "description": "Student not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT Anda di sini. Contoh: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}